feat: add Redis caching support
- Add Redis to docker-compose.yml
- Create cache module with get/set/delete/invalidate operations
- Add AppState to manage cache connection
- Connect to Redis on application startup via fairing
- Update Server model to use i64 for PostgreSQL BIGSERIAL
- Update .env.example with REDIS_URL

+++ src/db.rs
@@ -2,10 +2,11 @@ use crate::schema::servers;
 use diesel::pg::PgConnection;
 use diesel::prelude::*;
 
-#[derive(Queryable, Debug, serde::Serialize)]
+#[derive(Queryable, Selectable, Debug, serde::Serialize)]
+#[diesel(table_name = servers)]
 #[serde(rename_all = "camelCase")]
 pub struct Server {
-    pub id: i32,
+    pub id: i64,
     pub domain: String,
     pub name: Option<String>,
     pub description: Option<String>,
